Remove implicit conversion warning

+++ src/math/floatconvert.c
@@ -508,7 +508,7 @@ _extractexp(
 
   logbase = lgbase(base);
   decprec = DECPRECISION + 3;
-  exp = aprxlog10fn(x) * 3.321928095f;
+  exp = (int)(aprxlog10fn(x) * 3.321928095f);
   if (float_getexponent(x) < 0)
     exp -= 3;
   exp /= logbase;
